Local fork of angular-odata implementation
I chose this example project to work with Angular and OData, but unfortunately the implementation I chose is not yet up-to-date to build with Angular 5. To work around that quickly, I forked the handfuls of files into this project. A future update to the upstream project should let the stuff go away.