v0.9.13, Multi-column data, CSS example
Remove the old "tooltip" support.

Generate demo data with multiple columns.

Pass-through the richer data so that is available to templates.

Create example company selector which uses brutal CSS approach to create a multicolumn drop-down.

(It is not clear whether it is possible with angular material to put a multicolumn grid component in an auto complete; or whether that would even meet the material spec.)