v0.10.0, replace pair of functions with DataSource concept
Previous to this change, the component required that the user provide two functions via separate bindings. These functions themselves needed to be bound to the appropriate "this", and together this made for a very un-idiomatic Angular developer experience.

With this change, the application developer now provides an object which implements a new DataSource interface; this is much more similar both conceptually and syntactically to the mechanism for providing a data source to (for example) a mat-grid.

All demos updated;

Application code using previous versions will need to be updated similarly.