v0.11.1, add optional match function
This new match function implementation provides the capability that the older match field idea had been looking for. It is now easy for an application using this component/library to enable (for example) case insensitive matching of values typed by a user. There is an example provided.

Also, upon further testing the setTimeout() previously used by the component implementation turns out to not be necessary. The correct behavior (described by newly added comments) is achieved without it.

TODO: Find the right way to add unit and E2E tests for this overall functionality.