v0.11.2, don't debounce choosing from the list
Before this change, choosing from the list would propagate the new choice to the containing form after the debounce period. It's not clear whether it is "legal" in an Angular form for a non-focused control to emit a value... and I had a report of trouble when the control is used quickly. The change avoids that problem, propagating the selection immediately, while still leaving the "legal" question unanswered.