Rename to <search-select> @oasisdigital/angular-material-search-select
The old name, "observable autocomplete", was both distracting and wrong. The control provided uses a material auto complete under the hood, but what it provides is more typically called a "search select" or "select with search". Almost everything in Angular uses observable in some way so that word didn't add much information either.

There is no functional change in this commit, but it touches many many files; so there is likely to be some follow-up loose and work afterward.